rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.0",
    "langchain>=0.3.25",
    "ollama>=0.5.1",
    "pydantic>=2.11.5",
    "weaviate-client>=4.15.0",
]
